+++ ots/source/summons.h
@@ -7,7 +7,7 @@
 // modify it under the terms of the GNU General Public License
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -31,7 +31,7 @@ private:
 	static SummonMap summons;
 public:
 	static bool Load();
-	static int getRequiredMana(std::string name);
+	static int64_t getRequiredMana(std::string name);
 };
 
 #endif //SUMMONS_H
